ai_services/voice_analysis/requirements.txt
openai-whisper
numpy>=1.24.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.1
torch>=2.0.0
ffmpeg-python
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
